Modify the files so that each tool is in a separate file within the tools folder.
