+++ src/logcat/adb.ts
